requirements.txt
Flask==1.1.1
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
json5==0.8.5 
Markdown==3.1.1
Pillow==6.2.1
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
pytesseract==0.3.7
numpy
libopencv
opencv_python